TST: einsum: Move einsum tests to a different file
